+++ packages/ssr/src/lib/safe-stringify.js
@@ -1,0 +1,1 @@
+module.exports = input => JSON.stringify(input).replace(/</g, "\\u003c");