+++ packages/author-profile/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="4.1.5"></a>
+## [4.1.5](https://github.com/newsuk/times-components/compare/@times-components/author-profile@4.1.4...@times-components/author-profile@4.1.5) (2018-10-25)
+
+**Note:** Version bump only for package @times-components/author-profile
+
+
+
+
+
 <a name="4.1.4"></a>
 ## [4.1.4](https://github.com/newsuk/times-components/compare/@times-components/author-profile@4.1.2...@times-components/author-profile@4.1.4) (2018-10-23)