+++ packages/related-articles/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="4.3.4"></a>
+## [4.3.4](https://github.com/newsuk/times-components/compare/@times-components/related-articles@4.3.3...@times-components/related-articles@4.3.4) (2018-10-25)
+
+**Note:** Version bump only for package @times-components/related-articles
+
+
+
+
+
 <a name="4.3.3"></a>
 ## [4.3.3](https://github.com/newsuk/times-components/compare/@times-components/related-articles@4.3.1...@times-components/related-articles@4.3.3) (2018-10-23)