+++ packages/topic/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="3.4.5"></a>
+## [3.4.5](https://github.com/newsuk/times-components/compare/@times-components/topic@3.4.4...@times-components/topic@3.4.5) (2018-10-25)
+
+**Note:** Version bump only for package @times-components/topic
+
+
+
+
+
 <a name="3.4.4"></a>
 ## [3.4.4](https://github.com/newsuk/times-components/compare/@times-components/topic@3.4.2...@times-components/topic@3.4.4) (2018-10-23)